refactor: inline
