+++ packages/agent-runtime/src/server.ts
@@ -3,16 +3,24 @@ import { spawnSync } from "node:child_process";
 import { fileURLToPath } from "node:url";
 import { dirname, join } from "node:path";
 import { readFileSync, writeFileSync, existsSync } from "node:fs";
-import { WebSocketServer } from "ws";
+import headlessPkg from "@xterm/headless";
+const { Terminal: HeadlessTerminal } = headlessPkg;
+import serializePkg from "@xterm/addon-serialize";
+const { SerializeAddon } = serializePkg;
+import * as nodePty from "@lydell/node-pty";
+import { WebSocketServer, type WebSocket as WsWebSocket } from "ws";
 import { createHTTPHandler } from "@trpc/server/adapters/standalone";
 import { appRouter } from "agent-runtime-api/router";
 import type { AgentRuntimeContext } from "agent-runtime-api";
+import {
+  OP_INPUT, OP_OUTPUT, OP_RESIZE,
+  decodeFrame, encodeDataFrame, encodeExit,
+} from "api-server-api";
 import { createFilesService } from "./modules/files.js";
 import { composeSkills } from "./modules/skills/index.js";
 import { config } from "./modules/config.js";
 import { composeAcp } from "./modules/acp/compose.js";
 import { createWebSocketChannel } from "./modules/acp/infrastructure/create-websocket-channel.js";
-import { createPtyManager } from "./modules/pty/pty-session.js";
 import { startTriggerWatcher, type TriggerWatcher } from "./trigger-watcher.js";
 import { startPodFilesSync } from "./modules/pod-files/index.js";
 
@@ -65,11 +73,157 @@ const { runtime: acpRuntime } = composeAcp({
   log: (msg) => process.stderr.write(`[acp] ${msg}\n`),
 });
 
-const ptyManager = createPtyManager({
-  command: ["/usr/local/bin/harness-terminal"],
-  workingDir: workDir,
-  log: (msg) => process.stderr.write(`[pty] ${msg}\n`),
-});
+// ── PTY state for terminal-mode sessions ──────────────────────────────────────
+// One PTY per session id, kept alive across client reconnects via a headless
+// xterm that mirrors output and serializes the screen for replay.
+const PTY_DETACH_GRACE_MS = 30_000;
+
+interface PtySlot {
+  pty: nodePty.IPty | null;
+  headless: InstanceType<typeof HeadlessTerminal>;
+  serialize: InstanceType<typeof SerializeAddon>;
+  client: WsWebSocket | null;
+  graceTimer: ReturnType<typeof setTimeout> | null;
+}
+
+const ptySlots = new Map<string, PtySlot>();
+const ptyLog = (msg: string) => process.stderr.write(`[pty] ${msg}\n`);
+
+function killPtySlot(sessionId: string): void {
+  const slot = ptySlots.get(sessionId);
+  if (!slot) return;
+  if (slot.graceTimer) clearTimeout(slot.graceTimer);
+  if (slot.pty) {
+    try { slot.pty.kill(); } catch {}
+  }
+  slot.headless.dispose();
+  ptySlots.delete(sessionId);
+  ptyLog(`[${sessionId}] PTY killed`);
+}
+
+function spawnPty(sessionId: string, cols: number, rows: number): nodePty.IPty {
+  const cleanEnv: Record<string, string> = {};
+  for (const [k, v] of Object.entries(process.env)) {
+    if (v !== undefined && !k.startsWith("npm_config_") && !k.startsWith("npm_lifecycle_"))
+      cleanEnv[k] = v;
+  }
+  cleanEnv.TERM = "xterm-256color";
+  cleanEnv.COLORTERM = "truecolor";
+  cleanEnv.HARNESS_SESSION_ID = sessionId;
+
+  return nodePty.spawn("/usr/local/bin/harness-terminal", [], {
+    name: "xterm-256color",
+    cols,
+    rows,
+    cwd: workDir,
+    env: cleanEnv,
+  });
+}
+
+function wirePty(sessionId: string, slot: PtySlot, p: nodePty.IPty): void {
+  p.onData((data) => {
+    slot.headless.write(data);
+    if (slot.client?.readyState === 1) {
+      slot.client.send(encodeDataFrame(OP_OUTPUT, data));
+    }
+  });
+  p.onExit(({ exitCode }) => {
+    ptyLog(`[${sessionId}] PTY exited with code ${exitCode}`);
+    if (slot.client?.readyState === 1) {
+      slot.client.send(encodeExit(exitCode));
+      slot.client.close(1000, "pty exited");
+    }
+    slot.pty = null;
+    slot.headless.dispose();
+    ptySlots.delete(sessionId);
+  });
+}
+
+function wirePtyClose(sessionId: string, slot: PtySlot, ws: WsWebSocket): void {
+  ws.on("close", () => {
+    if (slot.client === ws) {
+      slot.client = null;
+      if (slot.pty) {
+        ptyLog(`[${sessionId}] Client detached — starting grace timer`);
+        slot.graceTimer = setTimeout(() => {
+          ptyLog(`[${sessionId}] Grace expired — killing PTY`);
+          killPtySlot(sessionId);
+        }, PTY_DETACH_GRACE_MS);
+      }
+    }
+  });
+  ws.on("error", () => {
+    if (slot.client === ws) slot.client = null;
+  });
+}
+
+function attachPty(sessionId: string, ws: WsWebSocket, opts: { reset: boolean }): void {
+  if (opts.reset) killPtySlot(sessionId);
+  let initialized = false;
+  ws.binaryType = "nodebuffer";
+
+  ws.on("message", (raw: Buffer) => {
+    let frame;
+    try { frame = decodeFrame(raw); } catch { return; }
+
+    // Wait for the first RESIZE to know the client's terminal size before spawning.
+    if (!initialized && frame.op === OP_RESIZE) {
+      initialized = true;
+      const { cols, rows } = frame;
+
+      const existing = ptySlots.get(sessionId);
+      if (existing) {
+        if (existing.graceTimer) clearTimeout(existing.graceTimer);
+        if (existing.client && existing.client !== ws && existing.client.readyState === 1) {
+          ptyLog(`[${sessionId}] Replacing existing client`);
+          existing.client.close(1000, "replaced by new connection");
+        }
+        existing.client = ws;
+        existing.headless.resize(cols, rows);
+        if (existing.pty) existing.pty.resize(cols, rows);
+
+        ptyLog(`[${sessionId}] Reconnect — replaying buffer`);
+        try {
+          const serialized = existing.serialize.serialize();
+          ptyLog(`[${sessionId}] Replay buffer: ${serialized.length} chars`);
+          if (serialized.length > 0) ws.send(encodeDataFrame(OP_OUTPUT, serialized));
+        } catch (err) {
+          ptyLog(`[${sessionId}] Replay failed: ${(err as Error).message}`);
+        }
+        wirePtyClose(sessionId, existing, ws);
+        return;
+      }
+
+      const headless = new HeadlessTerminal({ cols, rows, scrollback: 1000, allowProposedApi: true });
+      const serialize = new SerializeAddon();
+      headless.loadAddon(serialize);
+      const slot: PtySlot = { pty: null, headless, serialize, client: ws, graceTimer: null };
+      ptySlots.set(sessionId, slot);
+      slot.pty = spawnPty(sessionId, cols, rows);
+      wirePty(sessionId, slot, slot.pty);
+      wirePtyClose(sessionId, slot, ws);
+      ptyLog(`[${sessionId}] Spawned PTY (${cols}x${rows})`);
+      return;
+    }
+
+    const slot = ptySlots.get(sessionId);
+    if (!slot) return;
+
+    switch (frame.op) {
+      case OP_INPUT:
+        if (slot.pty) slot.pty.write(new TextDecoder().decode(frame.data));
+        break;
+      case OP_RESIZE:
+        slot.headless.resize(frame.cols, frame.rows);
+        if (slot.pty) slot.pty.resize(frame.cols, frame.rows);
+        break;
+    }
+  });
+
+  ws.on("error", () => {});
+}
+
+// ── HTTP + WebSocket server ───────────────────────────────────────────────────
 
 const server = http.createServer((req, res) => {
   if (req.method === "OPTIONS") {
@@ -90,7 +244,7 @@ const server = http.createServer((req, res) => {
       queuedPrompts: s.queuedPromptCount,
       agentAlive: s.agentAlive,
       activeTriggers: triggerWatcher?.activeCount() ?? 0,
-      terminalActive: ptyManager.activeCount() > 0,
+      terminalActive: ptySlots.size > 0,
     };
     res.writeHead(200, { "Content-Type": "application/json", ...CORS }).end(JSON.stringify(status));
     return;
@@ -113,20 +267,14 @@ acpWss.on("connection", (ws) => {
   acpRuntime.attach(createWebSocketChannel(ws));
 });
 
-// Terminal connections are handled in the upgrade handler below
-// where we extract the sessionId from the query string.
-
 server.on("upgrade", (req, socket, head) => {
-  const { pathname } = new URL(req.url!, `http://${req.headers.host}`);
-  if (pathname === "/api/acp") {
+  const url = new URL(req.url!, `http://${req.headers.host}`);
+  if (url.pathname === "/api/acp") {
     acpWss.handleUpgrade(req, socket, head, (ws) => acpWss.emit("connection", ws, req));
-  } else if (pathname === "/api/terminal") {
-    const params = new URL(req.url!, `http://${req.headers.host}`).searchParams;
-    const sessionId = params.get("sessionId") ?? "default";
-    const reset = params.get("reset") === "1";
-    termWss.handleUpgrade(req, socket, head, (ws) => {
-      ptyManager.attach(sessionId, ws, { reset });
-    });
+  } else if (url.pathname === "/api/terminal") {
+    const sessionId = url.searchParams.get("sessionId") ?? "default";
+    const reset = url.searchParams.get("reset") === "1";
+    termWss.handleUpgrade(req, socket, head, (ws) => attachPty(sessionId, ws, { reset }));
   } else {
     socket.destroy();
   }
